+++ plugins/events/SimEventsPlugin.hh
@@ -28,6 +28,7 @@
 
 namespace gazebo
 {
+
   class SimEventsPlugin : public WorldPlugin
   {
     /// \brief Destrutor